requirements.txt
anyio==3.6.2
asgiref==3.5.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
cachetools==5.3.0
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.1.0
cloudinary==1.30.0
confusable-homoglyphs==3.2.0
cryptography==37.0.4
cssselect==1.2.0
cssutils==2.6.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.1.2
django-allauth==0.51.0
django-appconf==1.0.5
django-bootstrap-v5==1.0.11
django-cloudinary-storage==0.3.0
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-heroku==0.3.1
django-imagekit==4.1.0
django-mpesa==2.0.14
django-on-heroku==1.1.2
django-registration==3.3
django-simple-mail==2.3.5
django-tinymce==3.4.0
djangorestframework==3.14.0
djangorestframework-pagination==1.1
djangorestframework-simplejwt==5.2.2
dnspython==2.3.0
email-validator==1.3.1
gunicorn==20.1.0
h11==0.14.0
html2text==2020.1.16
httpcore==0.16.3
httpx==0.23.3
idna==3.4
lxml==4.9.2
oauthlib==3.2.0
pilkit==2.0
Pillow==9.1.1
premailer==3.10.0
psycopg2==2.9.3
psycopg2-binary==2.9.5
pycodestyle==2.7.0
pycparser==2.21
pydantic==1.10.4
PyJWT==2.4.0
python-dateutil==2.8.2
python-decouple==3.5
python3-openid==3.2.0
pytz==2022.6
pyuploadcare==3.1.0
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.2.1
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.4.0
urllib3==1.26.7
whitenoise==6.2.0
zipp==3.6.0
